moved tests to test folder to match other micro app
