+++ apps/sodo-search/vite.config.mjs
@@ -28,7 +28,7 @@ export default defineConfig((config) => {
         ],
         esbuild: {
             loader: 'jsx',
-            include: /src\/.*\.jsx?$/,
+            include: /(src|test)\/.*\.jsx?$/,
             exclude: []
         },
         optimizeDeps: {
@@ -37,7 +37,7 @@ export default defineConfig((config) => {
                     {
                         name: 'load-js-files-as-jsx',
                         setup(build) {
-                            build.onLoad({filter: /src\/.*\.js$/}, async args => ({
+                            build.onLoad({filter: /(src|test)\/.*\.js$/}, async args => ({
                                 loader: 'jsx',
                                 contents: await fs.readFile(args.path, 'utf8')
                             }));
@@ -68,7 +68,7 @@ export default defineConfig((config) => {
         test: {
             globals: true,
             environment: 'jsdom',
-            setupFiles: './src/setup-tests.js',
+            setupFiles: './test/setup-tests.js',
             testTimeout: 10000
         }
     };
